geo transform caching for raycasts
now the geo passed to OpenCL to perform line casts, is only regenerated when a mesh has rotated or moved, or it's parent has. this means only a small amount to none cl geo data needs to be rebuilt per frame.
further optimizations to come. as raycasting is obviously a huge part of any 3D project.